refactor(gateway): modularize Codex bridge phase 2
