+++ gateway/codex/__init__.py
@@ -1,0 +1,1 @@
+"""Internal modules for the Codex-first Gateway bridge."""
